refactor: split engine module
